iter12 bigint

+++ internal/repository/metricRepository.go
@@ -114,7 +114,7 @@ func InitTables(ctx context.Context, db *sql.DB) {
             	id SERIAL PRIMARY KEY,
                 name VARCHAR(255) NOT NULL,
                 type_metrics VARCHAR(255) NOT NULL,
-            	delta INTEGER,
+            	delta bigint,
                 value DOUBLE PRECISION
 )
 `
